requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary]==3.2.11
sqlalchemy==2.0.36
pydantic-settings==2.1.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.1

